requirements.txt
streamlit
youtube-transcript-api
openai-whisper
deep-translator
yt-dlp
torch